requirements.txt
Django==5.2.7
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
drf-spectacular-sidecar==2024.5.1
python-dotenv==1.0.1
Pillow==10.4.0
psycopg2-binary==2.9.9
django-extensions==3.2.3
openpyxl==3.1.5
